Repository maintenance.
- Moved all scripts to `scripts` folder
- Dependencies are now defined in `setup.cfg` instead of conda/pip
   requirements.txt
- Dev dependencies are now defined in `setup.cfg`
- Updated contribution guidelines and installation instructions
- Switched test runner from `green` to `pytest`
- Updated GitHub actions runner

+++ scripts/generate_docs.sh
@@ -3,6 +3,13 @@
 # exit when any command fails
 set -e
 
+if [[ ! -f setup.cfg ]] || [[ ! -d neurox ]] || [[ ! -d scripts ]] || [[ ! -d tests ]]
+then
+  echo "generate_docs.sh must be run from the root of the repository"
+  exit 1
+fi
+
+
 rm -rf docs-generator
 sphinx-apidoc -o docs-generator neurox sphinx-apidoc --full -H 'NeuroX toolkit' -A 'Qatar Computing Research Institute, HBKU' --templatedir docs/sphinx_templates
 
@@ -11,6 +18,7 @@ mkdir -p _static/css
 mkdir -p _static/img
 
 cp -r ../docs/sphinx_templates/*.rst .
+cp ../docs/sphinx_templates/custom.css _static/css/custom.css
 cp ../logo.svg _static/img/logo.svg
 
 echo "
@@ -32,10 +40,13 @@ html_theme_options = {
     'use_repository_button': True,
     'use_issues_button': True,
     'use_fullscreen_button': False,
-    'home_page_in_toc': True
+    'home_page_in_toc': True,
+    # 'extra_navbar': '<div class=\"version-footer\">Select version: </div>',
 }
 
 autodoc_member_order = 'bysource'
+
+html_css_files = ['css/custom.css']
 " >> conf.py
 
 make html
